backend/requirements.txt
Flask==2.1.2
flask_bcrypt==1.0.1
flask_login==0.6.1
flask_sqlalchemy==2.5.1
Pillow==9.1.1
